% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CfFuncs.R
\name{adjust_disc}
\alias{adjust_disc}
\title{Adjusts the discount factors by a spread}
\usage{
adjust_disc(fd, spread)
}
\arguments{
\item{fd}{vector of discount factors used to discount cashflows in \code{1:length(fd)} periods}

\item{spread}{effective spread}
}
\description{
Adjusts the discount factors by a spread
}
\examples{
adjust_disc(fd = c(0.99, 0.98), spread = 0.01)
}
