\name{adjustDisc}
\alias{adjustDisc}
\title{Adjusts the discount factors by a spread}
\usage{
adjustDisc(fd, spread)
}
\arguments{
  \item{fd}{vector of discount factors used to discount
  cashflows in \code{1:length(fd)} periods}

  \item{effective}{spread}
}
\description{
Adjusts the discount factors by a spread
}

