\name{bulletCashflow}
\alias{bulletCashflow}
\title{Cashflow for a bullet loan}
\usage{
bulletCashflow(rate, maturity)
}
\arguments{
  \item{rate}{The rate of the loan, as an effective
  periodic rate}

  \item{maturity}{The maturity of the loan}
}
\description{
A bullet loan pays interest periodically and returns its
principal at maturity
}

