% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CfFuncs.R
\name{cashflow}
\alias{cashflow}
\title{Get the cashflow for a loan}
\usage{
cashflow(l)
}
\arguments{
\item{l}{The loan}
}
\description{
Returns the cashflow for the loan, excluding the initial inflow for the loan taker
}
\examples{
l <- loan(rate = 0.05, maturity = 10, amt = 100, type = "bullet")
cashflow(l)
}

