\name{frenchCashflow}
\alias{frenchCashflow}
\title{Cashflow for a french loan}
\usage{
frenchCashflow(rate, maturity)
}
\arguments{
  \item{rate}{The rate of the loan, as an effective
  periodic rate}

  \item{maturity}{The maturity of the loan}
}
\description{
A french loan has a constant payment
}

