\name{germanCashflow}
\alias{germanCashflow}
\title{Cashflow for a german loan}
\usage{
germanCashflow(rate, maturity)
}
\arguments{
  \item{rate}{The rate of the loan, as an effective
  periodic rate}

  \item{maturity}{The maturity of the loan}
}
\description{
A german loan has constant amortizations
}

