\name{get_rate_fun}
\alias{get_rate_fun}
\title{Returns a function f such that f(x) is the rate of type rate_type for time x}
\usage{
get_rate_fun(r, rate_type = "spot")
}
\arguments{
  \item{r}{A rate curve object}

  \item{type}{The rate type}
}
\description{
Returns a function f such that f(x) is the rate of type
rate_type for time x
}

