\name{getPaymt}
\alias{getPaymt}
\title{The value of the payment of a loan with constant payments (french type amortization)}
\usage{
getPaymt(amt, maturity, rate)
}
\arguments{
  \item{amt}{The amount of the loan}

  \item{maturity}{The maturity of the loan}

  \item{rate}{The rate of the loan}
}
\description{
The periodicity of the maturity and the rate must match,
and this will be the periodicity of the payments
}

