% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipmt.R
\name{ipmt}
\alias{ipmt}
\title{Interest portion of annuity payment}
\source{
builds heavily on \url{https://gist.github.com/raadk/dcd503815bbb271484ff}
}
\usage{
ipmt(irate, per, nper, pv, fv = 0, type = "immediate", silent = FALSE)
}
\arguments{
\item{irate}{Interest rate per period (see nper) as fraction (i.e. 5\% is .05 or 1/20)}

\item{per}{period of interest}

\item{nper}{Number of interest periods as integer (no fractions)}

\item{pv}{Present value}

\item{fv}{Future value (default is zero)}

\item{type}{type of the annuity: due or immediate?}

\item{silent}{Should warning messages be suppressed? Default is no.}
}
\value{
interest on annuity
}
\description{
Calculation of annuity interest.
}
\examples{
ipmt(.1, 1, 1, -1000) # a single period at 10\\% interest >> interest = 1/10 of principal

ipmt(.1, 1:5, 5, -1000) # see five payments of five; the amount of interest goes down

}
