% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmt.R
\name{ppmt}
\alias{ppmt}
\title{Principal portion of annuity payment}
\source{
builds heavily on \url{https://gist.github.com/raadk/dcd503815bbb271484ff}
}
\usage{
ppmt(irate, per, nper, pv, fv = 0, type = "immediate", silent = FALSE)
}
\arguments{
\item{irate}{Interest rate per period (see nper)}

\item{per}{period of interest}

\item{nper}{Number of interest periods}

\item{pv}{present value}

\item{fv}{future value (default is zero)}

\item{type}{type of the annuity: due or immediate?}

\item{silent}{Should warning messages be suppressed? Default is no.}
}
\value{
principal repayment
}
\description{
Calculation of principal repayment portion of an anuity payment.
}
\examples{

}
