% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CfFuncs.R
\name{xnpv}
\alias{xnpv}
\title{Net Present Value of an irregular cashflow (NPV)}
\usage{
xnpv(i, cf, d)
}
\arguments{
\item{i}{The rate used to discount the cashflow. It must be an effective anual rate (EAR)}

\item{cf}{The cashflow}

\item{d}{The dates when each cashflow occurs. Same length as the cashflow}
}
\description{
Net Present Value of an irregular cashflow (NPV)
}
\examples{
xnpv(i = 0.01, cf = c(-1, 0.5, 0.9), d = as.Date(c("2015-01-01", "2015-02-15", "2015-04-10")))
}

