\name{zeroCashflow}
\alias{zeroCashflow}
\title{Cashflow for a zero loan}
\usage{
zeroCashflow(rate, maturity)
}
\arguments{
  \item{rate}{The rate of the loan, as an effective rate
  with periodicity equal to the maturity of the loan}

  \item{maturity}{The maturity of the loan}
}
\description{
A zero loan makes an only payment, paying everything at
maturity
}

