% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.R
\name{dcf}
\alias{dcf}
\title{DCF – Day Count Fraction}
\usage{
dcf(
  date_from = Sys.Date(),
  date_to = Sys.Date(),
  convention = "ACT/365",
  silent = FALSE
)
}
\arguments{
\item{date_from}{Start date (or a vector of dates)}

\item{date_to}{End date (or a vector of dates)}

\item{convention}{Interest rate convention; default is "ACT/365".}

\item{silent}{Should warning messages be suppressed? Default is no.}
}
\value{
Day Count Fraction - difference in days divided by days per year
}
\description{
DCF – Day Count Fraction
}
\examples{
dcf(as.Date("2019-01-01"), as.Date("2019-01-31"), convention = "30/360")
}
