% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CfFuncs.R
\name{disc_cf}
\alias{disc_cf}
\title{Value of a discounted cashflow}
\usage{
disc_cf(fd, cf)
}
\arguments{
\item{fd}{The discount factor vector}

\item{cf}{The cashflow}
}
\description{
Value of a discounted cashflow
}
\examples{
disc_cf(fd = c(1, 0.99, 0.98, 0.97), cf = c(1, -0.3, -0.4, -0.6))
}

