% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr}
\alias{irr}
\title{IRR – Internal Rate of Return}
\usage{
irr(
  cashflows,
  dates,
  method = "compound",
  convention = "ACT/365",
  silent = FALSE
)
}
\arguments{
\item{cashflows}{Cash flows, as a vector.}

\item{dates}{Value dates for the cashflows vector.}

\item{method}{Compound or simple interest? Default is compound.}

\item{convention}{Interest rate convention; default is "ACT/365".}

\item{silent}{Should warning messages be suppressed? Default is no.}
}
\value{
Internal Rate of Return - as a fraction, per annum
}
\description{
Rate at which all cashflows will be discounted to exactly zero.
}
\examples{
irr(
  dates = as.Date(c("2019-01-01", "2020-01-01")),
  cashflows = c(-1000, 1100)
)
}
\seealso{
npv
}
