% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv.R
\name{npv}
\alias{npv}
\title{Net Present Value}
\usage{
npv(interest, cashflows, dates, convention = "ACT/365", silent = FALSE)
}
\arguments{
\item{interest}{Interest rate per annum}

\item{cashflows}{Cash flows, as a vector}

\item{dates}{Value dates for the cashflows vector}

\item{convention}{Interest rate convention; default is ACT/365}

\item{silent}{should warning messages be suppressed? Default is no.}
}
\value{
Net present value of future cash flows
}
\description{
Discounted value of future cash flows; the date of the first cash flow is considered as origin.
}
\examples{
npv(
  interest = 10 / 100,
  dates = as.Date(c("2019-01-01", "2020-01-01")),
  cashflows = c(0, 1100)
)
}
