% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CfFuncs.R
\name{pmt}
\alias{pmt}
\title{The value of the payment of a loan with constant payments (french type amortization)}
\usage{
pmt(amt, maturity, rate)
}
\arguments{
\item{amt}{The amount of the loan}

\item{maturity}{The maturity of the loan}

\item{rate}{The rate of the loan}
}
\description{
The value of the payment of a loan with constant payments (french type amortization)
}
\details{
The periodicity of the maturity and the rate must match, and this will be the periodicity of the payments
}
\examples{
pmt(amt = 100, maturity = 10, rate = 0.05)
}
