\name{rem}
\alias{rem}
\title{Remaining capital in a loan}
\usage{
rem(cf, amt, r)
}
\arguments{
  \item{cf}{The cashflow of the loan, not including the
  initial inflow for the loan taker}

  \item{amt}{The original amount of the loan}

  \item{r}{The periodic rate of the loan}
}
\description{
The amount that has to be repayed at each moment in a loan,
at the end of the period
}
\examples{
rem(cf = rep_len(0.4, 4), amt = 1, r = 0.2)
}

