% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvm-package.R
\docType{package}
\name{tvm-package}
\alias{tvm}
\alias{tvm-package}
\title{tvm: Time Value of Money Functions}
\description{
Functions for managing cashflows and interest rate curves.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/juancentro/tvm}
}

}
\keyword{internal}
