\name{tvm-package}
\alias{tvm-package}
\alias{tvm}
\docType{package}
\title{
Time Value Functions
}
\description{
Functions for managing cashflows and interest rate curves
}
\details{
  \tabular{ll}{
  Package: \tab tvm\cr
  Type: \tab Package\cr
  Version: \tab 0.1\cr
  Date: \tab 2013-12-26\cr
  License: \tab MIT\cr
  }
}
\author{
Juan Manuel Truppia

Maintainer: Juan Manuel Truppia <jmtruppia@gmail.com>
}

\keyword{ package }
