% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CfFuncs.R
\name{xirr}
\alias{xirr}
\title{The IRR is returned as an effective anual rate}
\usage{
xirr(cf, d, interval = c(-1, 10), ...)
}
\arguments{
\item{cf}{The cashflow}

\item{d}{The dates when each cashflow occurs. Same length as the cashflow}

\item{interval}{A length 2 vector that indicates the root finding algorithm where to search for the irr}

\item{...}{Other arguments to be passed on to uniroot}
}
\description{
Internal Rate of Return of an irregular cashflow (IRR)
}
\examples{
xirr(cf = c(-1, 1.5), d = Sys.Date() + c(0, 365))
}

